Updated Podspec so that it always points to the most recent tag
Previously the pod's version and the code's tag had to both be updated by hand. This lead to the situation where the pod's version was bumped to 1.3.0, a new tag was created for 1.3.0 but the source path wasn't updated to point to the new tag resulting is users getting 1.2 when they thought they had 1.3.0. Now when you change the pod's version it will search for a tag with a matching version number.